perf: 优化调整区块 Context 配置
