+++ FormStepForm/src/index.tsx
@@ -1,7 +1,7 @@
-import React, { useState, useEffect, useContext, useReducer } from 'react';
+import React, { useState, useEffect } from 'react';
 import { Card, Steps } from 'antd';
 import { PageHeaderWrapper } from '@ant-design/pro-layout';
-import { useModel, FormContext, StateType } from './useModel';
+import { useModel, FormStepProvider } from './context';
 import Step1 from './components/Step1';
 import Step2 from './components/Step2';
 import Step3 from './components/Step3';
@@ -21,10 +21,9 @@ const getCurrentStepAndComponent = (current?: string) => {
   }
 };
 
-const FormWrap: React.FC<{}> = () => {
+const FormStepContent: React.FC<{}> = () => {
   const [stepComponent, setStepComponent] = useState<React.ReactNode>(<Step1 />);
   const [currentStep, setCurrentStep] = useState<number>(0);
-
   const { current } = useModel();
 
   useEffect(() => {
@@ -49,24 +48,11 @@ const FormWrap: React.FC<{}> = () => {
   );
 };
 
-function reducer(state: StateType, { type, payload }: { type: string; payload?: any }) {
-  switch (type) {
-    case 'saveCurrentStep':
-      return { ...state, current: payload };
-    case 'saveStepFormData':
-      return { ...state, step: payload };
-    default:
-      throw new Error(`action ${type} does not exist!`);
-  }
-}
-
 const PAGE_NAME_UPPER_CAMEL_CASE: React.FC<{}> = () => {
-  const { state: initialState } = useContext(FormContext);
-  const [state, dispatch] = useReducer(reducer, initialState);
   return (
-    <FormContext.Provider value={{ state, dispatch }}>
-      <FormWrap />
-    </FormContext.Provider>
+    <FormStepProvider>
+      <FormStepContent />
+    </FormStepProvider>
   );
 };
