--- FormStepForm/src/useModel.ts
@@ -1,47 +1,0 @@
-import { createContext, useContext } from 'react';
-
-export interface StateType {
-  current: string;
-  step: {
-    payAccount: string;
-    receiverAccount: string;
-    receiverName: string;
-    amount: string;
-  };
-}
-
-interface FormContextType {
-  state: StateType;
-  dispatch: ({ type, payload }: { type: string; payload?: any }) => void;
-}
-
-const defaultValue: FormContextType = {
-  state: {
-    current: 'info',
-    step: {
-      payAccount: 'ant-design@alipay.com',
-      receiverAccount: 'test@example.com',
-      receiverName: 'Alex',
-      amount: '500',
-    },
-  },
-  dispatch: () => {},
-};
-// 创建 context
-export const FormContext = createContext<FormContextType>(defaultValue);
-
-export function useModel() {
-  const {
-    state: { current, step },
-    dispatch,
-  } = useContext(FormContext);
-
-  function setCurrent(curr: string) {
-    dispatch({ type: 'saveCurrentStep', payload: curr });
-  }
-  function setStepData(values: any) {
-    dispatch({ type: 'saveStepFormData', payload: values });
-  }
-
-  return { current, setCurrent, step, setStepData };
-}
